requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
pip==22.1.1
gunicorn==20.1.0
Flask==2.1.2
pandas==1.3.5
numpy==1.22.4
requests==2.27.1
Pillow==9.0.1
opencv-python-headless==4.5.5.64
torchsummary==1.5.1
torch==1.11.0+cpu
torchvision==0.12.0+cpu
Werkzeug==2.1.1